# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * THE POKER LOUNGE API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Player {
  id: number;
  firstName: string;
  lastName: string;
  phone: string;
  isGuest: boolean;
  createdAt: string;
}

export interface CreatePlayerBody {
  firstName: string;
  lastName: string;
  phone: string;
  isGuest?: boolean;
}

export interface PlayerStats {
  playerId: number;
  firstName: string;
  lastName: string;
  phone: string;
  totalGames: number;
  totalBuyins: number;
  totalProfit: number;
  winRate: number;
  biggestWin: number;
  biggestLoss: number;
}

export type SessionStatus = (typeof SessionStatus)[keyof typeof SessionStatus];

export const SessionStatus = {
  active: "active",
  closed: "closed",
} as const;

export interface Session {
  id: number;
  status: SessionStatus;
  startedAt: string;
  closedAt?: string | null;
  totalRake: number;
}

export type ActiveSessionStatus =
  (typeof ActiveSessionStatus)[keyof typeof ActiveSessionStatus];

export const ActiveSessionStatus = {
  active: "active",
  closed: "closed",
} as const;

export interface Buyin {
  id: number;
  sessionId: number;
  playerId: number;
  amount: number;
  chips: number;
  createdAt: string;
}

export interface SessionPlayerDetail {
  id: number;
  sessionId: number;
  playerId: number;
  totalBuyins: number;
  finalChips?: number | null;
  player: Player;
  buyins: Buyin[];
}

export interface ActiveSession {
  id: number;
  status: ActiveSessionStatus;
  startedAt: string;
  closedAt?: string | null;
  totalRake: number;
  players: SessionPlayerDetail[];
}

export interface SessionPlayer {
  id: number;
  sessionId: number;
  playerId: number;
  totalBuyins: number;
  finalChips?: number | null;
}

export interface CreateSessionBody {
  note?: string;
}

export interface AddPlayerToSessionBody {
  playerId: number;
  initialBuyin: number;
}

export interface RecordBuyinBody {
  playerId: number;
  amount: number;
}

export type CloseSessionBodyFinalChipsItem = {
  playerId: number;
  chips: number;
};

export interface CloseSessionBody {
  finalChips: CloseSessionBodyFinalChipsItem[];
}

export interface PlayerSettlement {
  playerId: number;
  firstName: string;
  lastName: string;
  phone: string;
  totalBuyins: number;
  finalChips: number;
  profit: number;
  rake: number;
  netProfit: number;
}

export interface Transfer {
  fromPlayerId: number;
  fromName: string;
  fromPhone: string;
  toPlayerId: number;
  toName: string;
  toPhone: string;
  amount: number;
}

export interface SettlementResult {
  sessionId: number;
  totalPot: number;
  totalRake: number;
  players: PlayerSettlement[];
  transfers: Transfer[];
}

export interface GroupBalance {
  totalRake: number;
  sessionsCount: number;
}

export type SearchPlayersParams = {
  q: string;
};