# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * THE POKER LOUNGE API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all players
 */
export const ListPlayersResponseItem = zod.object({
  id: zod.number(),
  firstName: zod.string(),
  lastName: zod.string(),
  phone: zod.string(),
  isGuest: zod.boolean(),
  createdAt: zod.string(),
});
export const ListPlayersResponse = zod.array(ListPlayersResponseItem);

/**
 * @summary Create a new player
 */
export const CreatePlayerBody = zod.object({
  firstName: zod.string(),
  lastName: zod.string(),
  phone: zod.string(),
  isGuest: zod.boolean().optional(),
});

/**
 * @summary Search players by name (autocomplete)
 */
export const SearchPlayersQueryParams = zod.object({
  q: zod.coerce.string(),
});

export const SearchPlayersResponseItem = zod.object({
  id: zod.number(),
  firstName: zod.string(),
  lastName: zod.string(),
  phone: zod.string(),
  isGuest: zod.boolean(),
  createdAt: zod.string(),
});
export const SearchPlayersResponse = zod.array(SearchPlayersResponseItem);

/**
 * @summary Get a player by ID
 */
export const GetPlayerParams = zod.object({
  id: zod.coerce.number(),
});

export const GetPlayerResponse = zod.object({
  id: zod.number(),
  firstName: zod.string(),
  lastName: zod.string(),
  phone: zod.string(),
  isGuest: zod.boolean(),
  createdAt: zod.string(),
});

/**
 * @summary Get lifetime stats for a player
 */
export const GetPlayerStatsParams = zod.object({
  id: zod.coerce.number(),
});

export const GetPlayerStatsResponse = zod.object({
  playerId: zod.number(),
  firstName: zod.string(),
  lastName: zod.string(),
  phone: zod.string(),
  totalGames: zod.number(),
  totalBuyins: zod.number(),
  totalProfit: zod.number(),
  winRate: zod.number(),
  biggestWin: zod.number(),
  biggestLoss: zod.number(),
});

/**
 * @summary List all sessions
 */
export const ListSessionsResponseItem = zod.object({
  id: zod.number(),
  status: zod.enum(["active", "closed"]),
  startedAt: zod.string(),
  closedAt: zod.string().nullish(),
  totalRake: zod.number(),
});
export const ListSessionsResponse = zod.array(ListSessionsResponseItem);

/**
 * @summary Start a new session
 */
export const CreateSessionBody = zod.object({
  note: zod.string().optional(),
});

/**
 * @summary Get the currently active session
 */
export const GetActiveSessionResponse = zod.object({
  id: zod.number(),
  status: zod.enum(["active", "closed"]),
  startedAt: zod.string(),
  closedAt: zod.string().nullish(),
  totalRake: zod.number(),
  players: zod.array(
    zod.object({
      id: zod.number(),
      sessionId: zod.number(),
      playerId: zod.number(),
      totalBuyins: zod.number(),
      finalChips: zod.number().nullish(),
      player: zod.object({
        id: zod.number(),
        firstName: zod.string(),
        lastName: zod.string(),
        phone: zod.string(),
        isGuest: zod.boolean(),
        createdAt: zod.string(),
      }),
      buyins: zod.array(
        zod.object({
          id: zod.number(),
          sessionId: zod.number(),
          playerId: zod.number(),
          amount: zod.number(),
          chips: zod.number(),
          createdAt: zod.string(),
        }),
      ),
    }),
  ),
});

/**
 * @summary Get a session by ID
 */
export const GetSessionParams = zod.object({
  id: zod.coerce.number(),
});

export const GetSessionResponse = zod.object({
  id: zod.number(),
  status: zod.enum(["active", "closed"]),
  startedAt: zod.string(),
  closedAt: zod.string().nullish(),
  totalRake: zod.number(),
  players: zod.array(
    zod.object({
      id: zod.number(),
      sessionId: zod.number(),
      playerId: zod.number(),
      totalBuyins: zod.number(),
      finalChips: zod.number().nullish(),
      player: zod.object({
        id: zod.number(),
        firstName: zod.string(),
        lastName: zod.string(),
        phone: zod.string(),
        isGuest: zod.boolean(),
        createdAt: zod.string(),
      }),
      buyins: zod.array(
        zod.object({
          id: zod.number(),
          sessionId: zod.number(),
          playerId: zod.number(),
          amount: zod.number(),
          chips: zod.number(),
          createdAt: zod.string(),
        }),
      ),
    }),
  ),
});

/**
 * @summary Close a session, record final chips, calculate settlement
 */
export const CloseSessionParams = zod.object({
  id: zod.coerce.number(),
});

export const CloseSessionBody = zod.object({
  finalChips: zod.array(
    zod.object({
      playerId: zod.number(),
      chips: zod.number(),
    }),
  ),
});

export const CloseSessionResponse = zod.object({
  sessionId: zod.number(),
  totalPot: zod.number(),
  totalRake: zod.number(),
  players: zod.array(
    zod.object({
      playerId: zod.number(),
      firstName: zod.string(),
      lastName: zod.string(),
      phone: zod.string(),
      totalBuyins: zod.number(),
      finalChips: zod.number(),
      profit: zod.number(),
      rake: zod.number(),
      netProfit: zod.number(),
    }),
  ),
  transfers: zod.array(
    zod.object({
      fromPlayerId: zod.number(),
      fromName: zod.string(),
      fromPhone: zod.string(),
      toPlayerId: zod.number(),
      toName: zod.string(),
      toPhone: zod.string(),
      amount: zod.number(),
    }),
  ),
});

/**
 * @summary Add a player to an active session
 */
export const AddPlayerToSessionParams = zod.object({
  id: zod.coerce.number(),
});

export const AddPlayerToSessionBody = zod.object({
  playerId: zod.number(),
  initialBuyin: zod.number(),
});

/**
 * @summary Record a buy-in for a player in a session
 */
export const RecordBuyinParams = zod.object({
  id: zod.coerce.number(),
});

export const RecordBuyinBody = zod.object({
  playerId: zod.number(),
  amount: zod.number(),
});

/**
 * @summary Get settlement details for a closed session
 */
export const GetSessionSettlementParams = zod.object({
  id: zod.coerce.number(),
});

export const GetSessionSettlementResponse = zod.object({
  sessionId: zod.number(),
  totalPot: zod.number(),
  totalRake: zod.number(),
  players: zod.array(
    zod.object({
      playerId: zod.number(),
      firstName: zod.string(),
      lastName: zod.string(),
      phone: zod.string(),
      totalBuyins: zod.number(),
      finalChips: zod.number(),
      profit: zod.number(),
      rake: zod.number(),
      netProfit: zod.number(),
    }),
  ),
  transfers: zod.array(
    zod.object({
      fromPlayerId: zod.number(),
      fromName: zod.string(),
      fromPhone: zod.string(),
      toPlayerId: zod.number(),
      toName: zod.string(),
      toPhone: zod.string(),
      amount: zod.number(),
    }),
  ),
});

/**
 * @summary Get the running group rake balance
 */
export const GetGroupBalanceResponse = zod.object({
  totalRake: zod.number(),
  sessionsCount: zod.number(),
});

/**
 * @summary Get all-time player leaderboard
 */
export const GetLeaderboardResponseItem = zod.object({
  playerId: zod.number(),
  firstName: zod.string(),
  lastName: zod.string(),
  phone: zod.string(),
  totalGames: zod.number(),
  totalBuyins: zod.number(),
  totalProfit: zod.number(),
  winRate: zod.number(),
  biggestWin: zod.number(),
  biggestLoss: zod.number(),
});
export const GetLeaderboardResponse = zod.array(GetLeaderboardResponseItem);